Resolve an issue with the systemd/sysvinit ansible conflict
Currently, there appears to be an issue with Ansible in which, if a
service is already enabled with sysvinit, the service will not be
enabled with Ansible. This appears to be the condition with this
repository; likely, the upstream package is bringing in the sysvinit
script (as it's not being deployed by ansible) and enabling it.

This commit implements a workaround as documented
https://github.com/ansible/ansible-modules-core/issues/3764#issuecomment-284331673.
While, confusingly, the package is "using" the sysvinit style service
tool for enabling the service, systemd implements a compatibility layer
that allows enabling it via this shim. This correctly enables the
service.

See https://github.com/ansible/ansible/issues/22303
See https://github.com/geerlingguy/ansible-role-php/commit/84ab337202415dbf2fe694f156128f721b502a84