Improve the code style in the main tasks.yml file
Recently, Ansible introduced the capability to sum "when" conditions as
a yaml list (something like a json array). When formatted thusly:

  when:
    - foo == "bar"
    - baz == "herp"

All conditionals are logically "and"ed together.

See http://docs.ansible.com/ansible/playbooks_conditionals.html#the-when-statement

Further, a line break was added to the bottom of this file.