Remove incorrect `use` param for non-xenial service
Previous work implemented a workaround for an ansible issue in which, in
some cases, Ansible would not detect correctly whether a service is
enabled. However, the author inadvertently committed a change to both
tasks.

This commit removes the change to the original task such that it will
use the ansible service auto-dectection, and not sysvinit